Finish login functionality
Implement functionality to check that a user trying to login exists in
the system, and then use bcrypt to compare the password entered to the
stored password to verify credentials.  Also implement serializing and
deserializing of user with passport.  Add passport middleware to app.js.
Make a global user variable and change the navbar based on if there is
an existing session or not.